+++ ProductRegistration/Modele/Dto/UserDto/SignupDto.cs
@@ -3,7 +3,7 @@ using Infrastructure.Entity.User;
 
 namespace ProductRegistration.Modele.Dto.UserDto;
 
-public class SignupDto:BaseDto<SignupDto,loginUser,int>
+public class SignupDto:BaseDto<SignupDto,loginUser,long>
 {
     public virtual string Fullname{ get; set; }
     public virtual string? UserName { get; set; }
